+++ src/app/main_tabs/diagrams/damage_resistance_chart.rs
@@ -4,11 +4,7 @@ use eframe::egui::*;
 use egui_plot::*;
 use itertools::Itertools;
 
-use crate::{
-    analyzer::*,
-    app::{settings::Settings, theme},
-    helpers::number_formatting::NumberFormatter,
-};
+use crate::{analyzer::*, app::settings::Settings, helpers::number_formatting::NumberFormatter};
 
 use super::common::*;
 
@@ -102,7 +98,7 @@ impl DamageResistanceChart {
 
         plot.show(ui, |p| {
             for (index, bars) in self.bars.iter().enumerate() {
-                p.bar_chart(bars.chart(settings).color(theme::series_color(index)));
+                p.bar_chart(bars.chart(settings).color(series_color(&bars.data, index)));
             }
         });
     }