+++ src/app/main_tabs/diagrams/values_chart.rs
@@ -2,10 +2,7 @@ use eframe::egui::*;
 use egui_plot::*;
 use itertools::Itertools;
 
-use crate::{
-    app::{settings::Settings, theme},
-    helpers::number_formatting::NumberFormatter,
-};
+use crate::{app::settings::Settings, helpers::number_formatting::NumberFormatter};
 
 use super::common::*;
 
@@ -110,7 +107,7 @@ impl<T: PreparedValue> ValuesChart<T> {
 
         plot.show(ui, |p| {
             for (index, bars) in self.bars.iter().enumerate() {
-                p.bar_chart(bars.chart(settings).color(theme::series_color(index)));
+                p.bar_chart(bars.chart(settings).color(series_color(&bars.data, index)));
             }
         });
     }